+++ garnet3d/main/src/priv/gfx/util/fakeRenderer.cpp
@@ -252,7 +252,7 @@ namespace GN { namespace gfx
 
         virtual bool drawBegin() { return true; }
         virtual void drawEnd() {}
-        virtual void clearScreen( const Vector4f & c, float z, UInt32 s, BitFields flags ) {}
+        virtual void clearScreen( const Vector4f & c, float z, UInt8 s, BitFields flags ) {}
         virtual void drawIndexed( PrimitiveType prim,
                                   size_t        numPrim,
                                   size_t        startvtx,
